requirements.txt
langchain==1.2.10
langchain-community==0.4.1
langchain-google-genai==4.2.1
langchain-classic==1.0.1
langgraph==1.0.10
neo4j==6.1.0
python-dotenv==1.2.2
